ml_service/requirements.txt
djangorestframework==3.14.0
django-environ==0.11.2